@urbit/http-api: fix package configuration
