--- pkg/npm/http-api/index.ts
@@ -1,3 +1,0 @@
-export * from './types';
-import Urbit from './Urbit';
-export { Urbit as default };
