+++ pkg/npm/http-api/src/types.ts
@@ -37,19 +37,20 @@ export interface UrbitInterface {
   sseClientInitialized: boolean;
   cookie?: string | undefined;
   outstandingPokes: Map<number, PokeHandlers>;
-  outstandingSubscriptions: Map<number, SubscriptionInterface>;
+  outstandingSubscriptions: Map<number, SubscriptionRequestInterface>;
   verbose?: boolean;
   ship?: string | null;
+  onError?: (error: any) => void;
   connect(): void;
   connect(): Promise<void>;
   eventSource(): void;
   getEventId(): number;
   ack(eventId: number): Promise<void | number>;
-  sendMessage(action: Action, data?: object): Promise<void | number>;
-  poke<T>(params: PokeInterface<T>): Promise<void | number>;
-  subscribe(params: SubscriptionRequestInterface): Promise<void | number>;
-  unsubscribe(subscription: string): Promise<void | number>;
-  delete(): Promise<void | number>;
+  // sendMessage(action: Action, data?: object): Promise<void | number>;
+  poke<T>(params: PokeInterface<T>): Promise<number>;
+  subscribe(params: SubscriptionRequestInterface): Promise<number>;
+  unsubscribe(subscription: number): Promise<boolean | void>;
+  delete(): void;
   scry(params: Scry): Promise<void | any>;
   thread<T>(params: Thread<T>): Promise<T>;
 }
@@ -63,4 +64,9 @@ export interface SSEOptions {
     Cookie?: string
   };
   withCredentials?: boolean;
+}
+
+export interface Message extends Record<string, any> {
+  action: Action;
+  id?: number;
 }
