+++ pkg/npm/http-api/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "include": ["*.ts"],
+  "include": ["src/**/*"],
   "exclude": ["node_modules", "dist", "@types"],
   "compilerOptions": {
 	  "outDir": "./dist",
